requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
distlib==0.3.9
Django==5.2.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
djangorestframework==3.16.0
filelock==3.18.0
Markdown==3.8
pillow==11.2.1
platformdirs==4.3.8
sqlparse==0.5.3
virtualenv==20.31.2
